backend/requirements.txt
# ==================== Core Framework ====================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.4.0
httpx>=0.27.0
flaml[automl]>=2.1.1

# ==================== Database & ORM ====================
sqlmodel>=0.0.22
sqlalchemy>=2.0.0
supabase==2.5.1

# ==================== Machine Learning ====================
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
joblib>=1.4.0

# ==================== Security & Authentication ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ==================== Configuration ====================
python-dotenv>=1.0.1

# ==================== Optional: For Production ====================
gunicorn>=22.0.0

# ==================== Optional: For Development ====================
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# flake8>=7.0.0
# mypy>=1.8.0